Config - Change income tax limit to REAL64 from 32.
- For consistency with existing interfaces. No functional changes.
- For Release Configuration, treat compilation warnings as errors.

+++ Server/Include/sp2_server.h
@@ -137,7 +137,7 @@ namespace SP2
       inline bool    DisbandAMDSOnOccupy() const { return m_bDisbandAMDSOnOccupy; }
       inline REAL32  GlobalTaxLimit() const { return m_fGlobalTaxLimit; }
       inline INT32   GlobalTaxSpecial(EGlobalTaxSpecialType::Enum in_eGlobalTaxSpecial) { return m_GlobalTaxSpecials.at(in_eGlobalTaxSpecial); }
-      inline REAL32  IncomeTaxLimit(EGovernmentType::Enum in_eGovernmentType) const { return m_IncomeTaxLimits.at(in_eGovernmentType); }
+      inline REAL64  IncomeTaxLimit(EGovernmentType::Enum in_eGovernmentType) const { return m_IncomeTaxLimits.at(in_eGovernmentType); }
       inline bool    LogBankruptcies() const { return m_bLogBankruptcies; }
       inline INT32   MaximumCellsInForeignCountry() const { return m_iMaximumCellsInForeignCountry; }
       inline GString Message() const { return m_sMessage; }
@@ -277,7 +277,7 @@ namespace SP2
       //! INT32 to eliminate precision issues with REAL32
       map<EGlobalTaxSpecialType::Enum, INT32> m_GlobalTaxSpecials;
 
-      map<EGovernmentType::Enum, REAL32> m_IncomeTaxLimits;
+      map<EGovernmentType::Enum, REAL64> m_IncomeTaxLimits;
 
       bool    m_bLogBankruptcies;
       INT32   m_iMaximumCellsInForeignCountry;
